Improve rounding logic
Instead of rounding everything to two decimal places, round based on the
number of lines in the file. If there are less than 500 lines, round to
the nearest integer. Less than 1000 lines rounds to the nearest tenths,
and more than that will be to nearest hundredths.

Everything will still display to two decimal points which is meh since
it would say, e.g. 24.00% which implies more precision than there
actually is, but it'll visually look better.

Bug: T200034