Compress test data to hide it from codesearch
These all use real MediaWiki names so they appear in codesearch results,
polluting the data. Since this is static data, let's convert it into a
binary format to obfuscate it from codesearch. gzip was picked since
it's ubiquitous enough and trivial to use on the command-line if you
want to peek at the actual files during development.

Bug: T338553
Change-Id: Ie7d67eedc7612a3618e8a5171268fa525f224918